requirements.txt
# Core ML / Audio
torch>=2.0.0
torchaudio>=2.0.0
torchcodec
speechbrain>=0.5.15

# Dataset
datasets>=2.14.0
huggingface_hub>=0.19.0

# ML
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Demo UI
gradio>=4.0.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0
